# CLN/PERF: remove gc.collect from the setting_with_copy checks as not needed