@jreback 2019-06-27
@pandas-dev/pandas-core ok with this?